Bugfix: avoid empty tooltip with multiple fields
This patch makes ToolTipControl behave the same way with multiple
fields like it does for a single field. If ToolTipControl is activated
for a VisualItem, which has other fields than the ToolTipControl
expects, it won't have any text to display in the tooltip. Therefore, it
should not even show the tooltip.