Bugfix: ranges on double and long values correctly mapped to int
When you work with a NumberRangeModel that is initialized with double
values, setting the value range with the method setValueRange() yields
an unexpected result (a tiny range close to the minimum is selected).
The bug is in updateRange(), where the double range is transformed to an
int range from 0 to 10,000. The problem is that the multiplication with
10,000 happens *after* the intermediate result is converted to int. So,
v and e are either 0 or 10,000.
The same bug and patch should apply to NumberRangeModel that are
initialized with long values. (I did not test it.)
NumberRangeModel that are initialized with int values are not affected
by this issue, because they do not need such an transformation.